+++ orchestration/temporal-java/src/test/java/edu/washu/tag/temporal/TemporalJavaApplicationTests.java
@@ -1,4 +1,4 @@
-package edu.washu.tag.temporaljava;
+package edu.washu.tag.temporal;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
